# Performance improvement - Introduce queue for raw statsd payloads