--- a/pkg/event/event.go
+++ b/pkg/event/event.go
@@ -118,7 +118,7 @@ func (eq *EventQueue) Flush() {
 
 func (eq *EventQueue) FlushUnlocked() {
 	eq.C <- eq.q
-	eq.q = make([]Event, 0, cap(eq.q))
+	eq.q = eq.q[:0]
 	eq.eventsFlushed.Inc()
 }