## Summary of this PR

In this PR I am introducing a queue for raw statsd payloads (calling them packets), and a worker pool to parse the packets in a parallel manner.

## Problem

I am currently managing a fleet of statsd exporters for a project and I noticed it suffers to keep up with high traffic applications. I have observed memory spikes that goes over 500Mib and when profiling those high memory usage instances I could see the hot path was mostly in statsd parsing ([see example heap profile](https://user-images.githubusercontent.com/2609731/185636434-20f2c89c-5544-4a3b-aab8-4443e12a5a00.png)).

I could not reach a conclusion on what exactly was causing the memory, I suspected at first it was the GC not being able to react fast enough to the burst of data or dangling references to the maps that get created while parsing tags.
Some things I have tried:

* Play with `GOGC` and try to make GC more aggressive: this yielded limited success, in general memory was lower, but we still saw spikes and OOM kills.
* Build the exporter with Go 1.19, set GOMEMLIMIT: almost same result as GOGC with more stability, but nevertheless OOM kills (with the limits we thing are acceptable for our case).
* To make sure it was a problem with the rate of which payloads get delivered to the exporter we tried rate-limiting with a custom tool before the exporter, this gave a good result, but metric correctness was hurt so we could not use it.


## Solution

Using knowledge of applications my team already built in the past and by looking at the code from the [Datadog Agent](https://github.com/DataDog/datadog-agent), I tried to introduce a "buffer" for statsd payloads *before* the exporter starts parsing the payloads. By doing this you free the "wire" of the UDP connection faster and allow for parallelism on the hard job that is parsing (hot path of heap allocation).

Apart from that, I moved the parsing from each listener to a pool of "workers" which are responsible for doing the work of parsing, relaying and send events to the EventQueue.


## Results

1. First of all **we don't have observed OOM Kills anymore**, even under higher loads (12 Mib/s of payload traffic).
2. **Memory usage** and spikes are **much lower**, also CPU throttle went away.
![image](https://user-images.githubusercontent.com/2609731/185640720-b5fed84e-42e6-4a3e-99db-128775b479c2.png)

### Bench comparison

I am omitting the benchmarks that are equal (parsing, line format, etc).

```
▶ benchstat /tmp/master-bench.txt /tmp/new-bench           
name                            old time/op    new time/op    delta
UDPListener1-8                    8.46µs ± 0%    3.39µs ± 0%   ~     (p=1.000 n=1+1)
UDPListener5-8                     179µs ± 0%     178µs ± 0%   ~     (p=1.000 n=1+1)
UDPListener50-8                   19.7ms ± 0%    17.7ms ± 0%   ~     (p=1.000 n=1+1)
ExporterListener-8                5.44ms ± 0%    6.07ms ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed1-8              5.45µs ± 0%    6.14µs ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed5-8              27.3µs ± 0%    30.9µs ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed50-8              277µs ± 0%     308µs ± 0%   ~     (p=1.000 n=1+1)
LineFormats/dogStatsd-8            390ns ± 0%     454ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidDogStatsd-8     468ns ± 0%     539ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/signalFx-8             411ns ± 0%     472ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidSignalFx-8      303ns ± 0%     348ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/influxDb-8             381ns ± 0%     453ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidInfluxDb-8      435ns ± 0%     529ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/statsd-8               254ns ± 0%     281ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidStatsd-8        392ns ± 0%     431ns ± 0%   ~     (p=1.000 n=1+1)

name                            old alloc/op   new alloc/op   delta
UDPListener1-8                    6.85kB ± 0%    5.25kB ± 0%   ~     (p=1.000 n=1+1)
UDPListener5-8                     171kB ± 0%     222kB ± 0%   ~     (p=1.000 n=1+1)
UDPListener50-8                   17.1MB ± 0%    22.3MB ± 0%   ~     (p=1.000 n=1+1)
ExporterListener-8                3.44MB ± 0%    3.44MB ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed1-8              4.56kB ± 0%    4.56kB ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed5-8              22.8kB ± 0%    22.8kB ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed50-8              228kB ± 0%     228kB ± 0%   ~     (p=1.000 n=1+1)
```


## Future TODOs

There are some things I would like to have, but I will leave them for another PR to avoid making this PR too big to review:

* I want to introduce metric for the size of both queues, we did that in other application by wrapping metric handler and passing the channels to them and calling `len()` before reporting metrics.