@pedro-stanaka 2022-08-22
So, seems like as soon you introduce parallelism on the workers (more than one routine parsing) we start getting problems with the race conditions. I will hold off this PR for now, if you want you can close. Once I have a proper solution I will reopen/open a new one.


Logs (if you might have a tip for me, I thank you)
```

ts=2022-08-22T08:55:57.155Z caller=main.go:296 level=info msg="Starting StatsD -> Prometheus Exporter" version="(version=, branch=, revision=)"
ts=2022-08-22T08:55:57.156Z caller=main.go:297 level=info msg="Build context" context="(go=go1.19, user=, date=)"
ts=2022-08-22T08:55:57.159Z caller=main.go:359 level=info msg="Accepting StatsD Traffic" udp=:9125 tcp= unixgram=
ts=2022-08-22T08:55:57.159Z caller=main.go:360 level=info msg="Accepting Prometheus Requests" addr=:9102
fatal error: concurrent map writes

goroutine 53 [running]:
github.com/prometheus/statsd_exporter/pkg/exporter.(*Exporter).handleEvent(0xc0000ace40, {0xa6e148, 0xc00054e1a0})
        /go/github.com/prometheus/statsd-exporter/pkg/exporter/exporter.go:109 +0x12f4
github.com/prometheus/statsd_exporter/pkg/exporter.(*Exporter).Listen(0xc0000ace40, 0xc0000aca20)
        /go/github.com/prometheus/statsd-exporter/pkg/exporter/exporter.go:72 +0x1ff
created by main.main
        /go/github.com/prometheus/statsd-exporter/main.go:546 +0x5d9e

goroutine 1 [select]:
main.main()
        /go/github.com/prometheus/statsd-exporter/main.go:552 +0x5e89

goroutine 26 [chan receive]:
github.com/prometheus/statsd_exporter/pkg/event.NewEventQueue.func1()
        /go/github.com/prometheus/statsd-exporter/pkg/event/event.go:94 +0x30
created by github.com/prometheus/statsd_exporter/pkg/event.NewEventQueue
        /go/github.com/prometheus/statsd-exporter/pkg/event/event.go:92 +0x1b9

goroutine 30 [runnable]:
net.wrapSyscallError(...)
        /usr/local/go/src/net/error_posix.go:17
net.(*netFD).readFromInet6(0xc0000bde00, {0xc00040ff69?, 0xde1dc0?, 0xc00040fe30?}, 0x46b70e?)
        /usr/local/go/src/net/fd_posix.go:74 +0x30
net.(*UDPConn).readFrom(0x0?, {0xc00040ff69?, 0xc00040ff18?, 0xc00040ff18?}, 0xc00041ff88)
        /usr/local/go/src/net/udpsock_posix.go:59 +0x85
net.(*UDPConn).readFromUDP(0xc0000b4548, {0xc00040ff69?, 0x8e?, 0xffff?}, 0xc00041ff88?)
        /usr/local/go/src/net/udpsock.go:149 +0x31
net.(*UDPConn).ReadFromUDP(...)
        /usr/local/go/src/net/udpsock.go:141
github.com/prometheus/statsd_exporter/pkg/listener.(*StatsDUDPListener).Listen(0xc00036ccc0)
        /go/github.com/prometheus/statsd-exporter/pkg/listener/listener.go:48 +0xe8
created by main.main
        /go/github.com/prometheus/statsd-exporter/main.go:395 +0x48dd

goroutine 31 [runnable]:
github.com/prometheus/statsd_exporter/pkg/line.buildEvent({0xc0001108f6, 0x2}, {0xc0001108cf, 0x1b}, 0x4079104db163baba, 0x0, 0xc0004406f0)
        /go/github.com/prometheus/statsd-exporter/pkg/line/line.go:64 +0x371
github.com/prometheus/statsd_exporter/pkg/line.(*Parser).LineToEvents(0x0?, {0xc0001108cf, 0x2f}, {0xc0000a00f0}, {0xa701a8, 0xc00006e6c0}, {0xa701a8, 0xc0002182a0}, {0xa701a8, 0xc00006e780}, ...)
        /go/github.com/prometheus/statsd-exporter/pkg/line/line.go:306 +0xa65
github.com/prometheus/statsd_exporter/pkg/parser.(*Worker).handle(0xc0000bde80, {0xc000110870?, 0x65?})
        /go/github.com/prometheus/statsd-exporter/pkg/parser/worker.go:84 +0x266
github.com/prometheus/statsd_exporter/pkg/parser.(*Worker).Consume(0xc0000bde80, 0x0?)
        /go/github.com/prometheus/statsd-exporter/pkg/parser/worker.go:72 +0x34
created by main.main
        /go/github.com/prometheus/statsd-exporter/main.go:490 +0x558c

goroutine 51 [IO wait]:
internal/poll.runtime_pollWait(0x7f603fcf8028, 0x72)
        /usr/local/go/src/runtime/netpoll.go:305 +0x89
internal/poll.(*pollDesc).wait(0xc000220000?, 0x6?, 0x0)
        /usr/local/go/src/internal/poll/fd_poll_runtime.go:84 +0x32
internal/poll.(*pollDesc).waitRead(...)
        /usr/local/go/src/internal/poll/fd_poll_runtime.go:89
internal/poll.(*FD).Accept(0xc000220000)
        /usr/local/go/src/internal/poll/fd_unix.go:614 +0x234
net.(*netFD).accept(0xc000220000)
        /usr/local/go/src/net/fd_unix.go:172 +0x35
net.(*TCPListener).accept(0xc000206300)
        /usr/local/go/src/net/tcpsock_posix.go:142 +0x28
net.(*TCPListener).Accept(0xc000206300)
        /usr/local/go/src/net/tcpsock.go:288 +0x3d
net/http.(*Server).Serve(0xc00021e000, {0xa6d508, 0xc000206300})
        /usr/local/go/src/net/http/server.go:3070 +0x385
net/http.(*Server).ListenAndServe(0xc00021e000)
        /usr/local/go/src/net/http/server.go:2999 +0x7d
net/http.ListenAndServe(...)
        /usr/local/go/src/net/http/server.go:3255
main.serveHTTP({0xa6a440?, 0xdb5700}, {0x7fffc5173454, 0x5}, {0xa69c00?, 0xc0000ae580?})
        /go/github.com/prometheus/statsd-exporter/main.go:172 +0xb8
created by main.main
        /go/github.com/prometheus/statsd-exporter/main.go:543 +0x5c7e

goroutine 52 [chan receive]:
main.sighupConfigReloader({0x7fffc517359f, 0x20}, 0x0?, {0xa69c00, 0xc0000ae580})
        /go/github.com/prometheus/statsd-exporter/main.go:180 +0x9a
created by main.main
        /go/github.com/prometheus/statsd-exporter/main.go:545 +0x5d3e

goroutine 55 [syscall]:
os/signal.signal_recv()
        /usr/local/go/src/runtime/sigqueue.go:152 +0x2f
os/signal.loop()
        /usr/local/go/src/os/signal/signal_unix.go:23 +0x19
created by os/signal.Notify.func1.1
        /usr/local/go/src/os/signal/signal.go:151 +0x2a

goroutine 345 [runnable]:
github.com/prometheus/statsd_exporter/pkg/mappercache/lru.(*metricMapperLRUCache).Add.func1()
        /go/github.com/prometheus/statsd-exporter/pkg/mappercache/lru/lru.go:53
runtime.goexit()
        /usr/local/go/src/runtime/asm_amd64.s:1594 +0x1
created by github.com/prometheus/statsd_exporter/pkg/mappercache/lru.(*metricMapperLRUCache).Add
        /go/github.com/prometheus/statsd-exporter/pkg/mappercache/lru/lru.go:53 +0x6e
```