@matthiasr 2022-08-28
One more thought because it's not immediately obvious (or I missed it) – what happens on configuration reload? Should we keep using the same workers, or replace and cancel them?