@matthiasr 2022-12-05
One thing that worries me about the parallelization is that we are re-introducing a channel that handles every packet. It's not quite as bad as the problems that #227 addressed as long as there's more than one line per packet, but in edge cases it might be.

On reflection, I am also worried about accidentally reordering events. That's probably OK™ for counter and histogram events but could cause wrong results with gauges. How could we prevent that without introducing more locking?