@kullanici0606 2023-09-18
While I was investigating an issue of k6 (https://github.com/grafana/k6/issues/2044), I also saw that UDP packets are dropped due to parsing and I also tried to fix the issue by separating reading UDP packets from socket and parsing them:

https://github.com/kullanici0606/statsd_exporter

(for diff: https://github.com/prometheus/statsd_exporter/compare/master...kullanici0606:statsd_exporter:master)

Then when I wanted to create a PR to get a feedback whether this was a good idea, I saw this PR, which has a very similar idea with mine.

The difference is that I only introduced one goroutine which means that packet ordering is not going to be an issue, but it may not be as performant as this PR.

Do you think that my changes can be used as a transitional way for this PR? 

If this PR is already on merging process, then I will not create another PR so that I do not take time of the developers of this repo.